moved PluginLoader classes to folder
