+++ CodeAnalyzerTool/CodeAnalyzerTool/PluginLoader/PluginLoadContext.cs
@@ -1,7 +1,7 @@
 ﻿using System.Reflection;
 using System.Runtime.Loader;
 
-namespace CodeAnalyzerTool;
+namespace CodeAnalyzerTool.PluginLoader;
 
 class PluginLoadContext : AssemblyLoadContext
 {
