+++ CodeAnalyzerTool/CodeAnalyzerTool/PluginLoader/PluginLoaderComposite.cs
@@ -1,7 +1,7 @@
 ﻿using CodeAnalyzerTool.Api;
 using CodeAnalyzerTool.Api.ConfigModel;
 
-namespace CodeAnalyzerTool;
+namespace CodeAnalyzerTool.PluginLoader;
 
 public class PluginLoaderComposite : PluginLoaderBase
 {
